Work on removing permissions

+++ Controller/OrganizationsController.php
@@ -264,10 +264,13 @@ class OrganizationsController extends AppController {
 		}
 		else 
 		{
-			return $this->redirect(array('supervisor' => true,
-										 'controller' => 'organizations',
-										 'action' => 'view',
-										 $id));
+			return $this->redirect(
+				array(
+						'supervisor' => true,
+						'controller' => 'organizations',
+						'action' => 'view',
+						$id)
+			);
 		}
 
 		$conditions = array(
@@ -289,15 +292,28 @@ class OrganizationsController extends AppController {
 			throw new NotFoundException(__('Invalid organization'));
 		}
 
-		$this->Organization->delete($id);
+		if(_CurrentUserCanWrite($this->Auth->user('user_id'))
+		{
+			$this->Organization->delete($id);
 
-		$this->redirect(
-			array(
-				'volunteer' => false,
-				'controller' => 'users',
-				'action' => 'activity'
-			)
-		);
+			$this->redirect(
+				array(
+					'volunteer' => false,
+					'controller' => 'users',
+					'action' => 'activity'
+				)
+			);
+		}
+		else
+		{
+			return $this->redirect(
+				array(
+					'volunteer' => false,
+					'controller' => $organizations,
+					'action' => 'index'
+				)
+			);
+		}
 	}
 
 
@@ -310,34 +326,47 @@ class OrganizationsController extends AppController {
 */
 	public function supervisor_view($organization_id = null) 
 	{
-		$sql_date_fmt = 'Y-m-d H:i:s';
+		if(_CurrentUserCanRead($this->Auth->user('user_id')))
+		{
+			$sql_date_fmt = 'Y-m-d H:i:s';
 
-		// summary all time
-		$users = $this->_GetUsersByOrganization($organization_id);
+			// summary all time
+			$users = $this->_GetUsersByOrganization($organization_id);
 
-		$event_ids = $this->Organization->Event->find('list', 
-			array(
-				'conditions' => array('Event.organization_id' => $organization_id),
-				'fields' => array('Event.event_id')
-			)
-		);
+			$event_ids = $this->Organization->Event->find('list', 
+				array(
+					'conditions' => array('Event.organization_id' => $organization_id),
+					'fields' => array('Event.event_id')
+				)
+			);
 
-		$events = $this->Organization->Event->find('all', array('conditions' => array('Event.event_id' => $event_ids)));
+			$events = $this->Organization->Event->find('all', array('conditions' => array('Event.event_id' => $event_ids)));
 
-		$conditions = array(
-			'Time.event_id' => $event_ids
-		);
-		$fields = array(
-			'User.*',
-			'SUM( TIMESTAMPDIFF(MINUTE, Time.start_time, Time.stop_time) )/60 as UserSumTime',
-			'COUNT( Time.time_id ) as UserNumberEvents'
-		);
-		$group = array(
-			'User.user_id'
-		);
+			$conditions = array(
+				'Time.event_id' => $event_ids
+			);
+			$fields = array(
+				'User.*',
+				'SUM( TIMESTAMPDIFF(MINUTE, Time.start_time, Time.stop_time) )/60 as UserSumTime',
+				'COUNT( Time.time_id ) as UserNumberEvents'
+			);
+			$group = array(
+				'User.user_id'
+			);
 
-		$userHours = $this->Organization->Event->Time->find('all', array('conditions' => $conditions, 'fields' => $fields, 'group' => $group) );
-		$this->set(compact('userHours', 'events'));
+			$userHours = $this->Organization->Event->Time->find('all', array('conditions' => $conditions, 'fields' => $fields, 'group' => $group) );
+			$this->set(compact('userHours', 'events'));
+		}
+		else
+		{
+			return $this->redirect(
+				array(
+					'volunteer' => false,
+					'controller' => $organizations,
+					'action' => 'index'
+				)
+			);
+		}
 	}
 
 
@@ -350,42 +379,55 @@ class OrganizationsController extends AppController {
 */
 	public function leave($organization_id = null) 
 	{
-		if ($organization_id != null) 
+		if($this->Auth->user('user_id') ) 
 		{
+			if ($organization_id != null) 
+			{
+				$conditions = array(
+					'Permission.user_id' => $this->Auth->user('user_id'),
+					'Permission.organization_id' => $organization_id
+				);
+				
+				if($this->Organization->Permission->deleteAll($conditions)) 
+				{
+					$this->Session->setFlash(__('Something went wrong. You cannot leave this organization.'), 'success');
+					$this->redirect(
+						array(
+							'volunteer' => false,
+							'controller' => 'users',
+							'action' => 'activity'
+						)
+					 );
+				}
+				else 
+				{
+					$this->Session->setFlash(__('Something went wrong. You cannot leave this organization.'), 'danger');
+				}
+			}
+
 			$conditions = array(
-				'Permission.user_id' => $this->Auth->user('user_id'),
-				'Permission.organization_id' => $organization_id
+				'Organization.organization_id' => $this->_GetUserOrganizationsByPermission('all')
 			);
-			
-			if($this->Organization->Permission->deleteAll($conditions)) 
-			{
-				$this->Session->setFlash(__('Something went wrong. You cannot leave this organization.'), 'success');
-				$this->redirect(
-					array(
-						'volunteer' => false,
-						'controller' => 'users',
-						'action' => 'activity'
-					)
-				 );
-			}
-			else 
-			{
-				$this->Session->setFlash(__('Something went wrong. You cannot leave this organization.'), 'danger');
-			}
+
+			$this->Paginator->settings = array(
+				'conditions' => $conditions,
+				'limit' => 10,
+				'contain' => array()
+			);
+
+			$data = $this->Paginator->paginate();
+			$this->set(compact('data'));
 		}
-
-		$conditions = array(
-			'Organization.organization_id' => $this->_GetUserOrganizationsByPermission('all')
-		);
-
-		$this->Paginator->settings = array(
-			'conditions' => $conditions,
-			'limit' => 10,
-			'contain' => array()
-		);
-
-		$data = $this->Paginator->paginate();
-		$this->set(compact('data'));
+		else
+		{
+			return $this->redirect(
+				array(
+					'volunteer' => false,
+					'controller' => $organizations,
+					'action' => 'index'
+				)
+			);
+		}
 	}
 
 
@@ -398,40 +440,183 @@ class OrganizationsController extends AppController {
 */
 	public function volunteer_leave($organization_id = null) 
 	{
-		if ($organization_id != null) 
+		if(_CurrentUserCanPublish($this->Auth->user('user_id')))
 		{
-			$entry['Permission'] = array(
-				'user_id' => $this->Auth->user('user_id'),
-				'organization_id' => $organization_id,
-				'publish' => false
+			if ($organization_id != null) 
+			{
+				$user_id = $this->Auth->user('user_id');
+				$permission = $this->Organization->Permission->findByUserIdAndOrganizationId($user_id, $organization_id); // yes, this will work
+				if( empty($permission) )
+				{
+				// permissions didn't exist.  redirect gracefully
+					$this->Session->setFlash(__('We are unable to process your request at this time.'), 'danger');
+				}
+				$this->Organization->Permission->id = $permission['Permission']['permission_id'];
+				if( $this->Organization->Permission->saveField('publish', false) )
+				{
+				// redirect
+					$this->Session->setFlash(__('You are no longer publishing your results to this organization.'), 'success');
+				}
+				else
+				{
+				// didn't save
+					$this->Session->setFlash(__('We are unable to process your request at this time.'), 'danger');
+				}
+			}
+
+			$conditions = array(
+				'Organization.organization_id' => $this->_GetUserOrganizationsByPermission('publish')
 			);
 
-			if($this->Organization->Permission->save($entry)) 
-			{
-				$this->Session->setFlash(__('You are no longer publishing your results to this organization.'), 'success');
-			}
-			else
-			{
-				$this->Session->setFlash(__('We are unable to process this request at this time.'), 'danger');
-			}
+			$fields = array(
+				'Organization.*',
+			);
+
+			$this->Paginator->settings = array(
+				'conditions' => $conditions,
+				'limit' => 10,
+				'contain' => array()
+			);
+
+			$data = $this->Paginator->paginate();
+			$this->set(compact('data'));
 		}
+		else 
+		{
+			return $this->redirect(
+				array(
+					'volunteer' => false,
+					'controller' => 'organizations',
+					'action' => 'leave'
+				)
+			);
+		}
+	}
 
-		$conditions = array(
-			'Organization.organization_id' => $this->_GetUserOrganizationsByPermission('publish')
-		);
 
-		$fields = array(
-			'Organization.*',
-		);
+/**
+ * volunteer_leave
+ *
+ * @throws ForbiddenException
+ * @param string $id
+ * @return void
+*/
+	public function coordinator_leave($organization_id = null) 
+	{
+		if(_CurrentUserCanRead($this->Auth->user('user_id')))
+		{
+			if ($organization_id != null) 
+			{
+				$user_id = $this->Auth->user('user_id');
+				$permission = $this->Organization->Permission->findByUserIdAndOrganizationId($user_id, $organization_id); // yes, this will work
+				if( empty($permission) )
+				{
+				// permissions didn't exist.  redirect gracefully
+					$this->Session->setFlash(__('We are unable to process your request at this time.'), 'danger');
+				}
+				$this->Organization->Permission->id = $permission['Permission']['permission_id'];
+				if( $this->Organization->Permission->saveField('write', false) )
+				{
+				// redirect
+					$this->Session->setFlash(__('You are no longer able to coordinate for this organization.'), 'success');
+				}
+				else
+				{
+				// didn't save
+					$this->Session->setFlash(__('We are unable to process your request at this time.'), 'danger');
+				}
+			}
 
-		$this->Paginator->settings = array(
-			'conditions' => $conditions,
-			'limit' => 10,
-			'contain' => array()
-		);
+			$conditions = array(
+				'Organization.organization_id' => $this->_GetUserOrganizationsByPermission('publish')
+			);
 
-		$data = $this->Paginator->paginate();
-		$this->set(compact('data'));
+			$fields = array(
+				'Organization.*',
+			);
+
+			$this->Paginator->settings = array(
+				'conditions' => $conditions,
+				'limit' => 10,
+				'contain' => array()
+			);
+
+			$data = $this->Paginator->paginate();
+			$this->set(compact('data'));
+		}
+		else 
+		{
+			return $this->redirect(
+				array(
+					'volunteer' => true,
+					'controller' => 'organizations',
+					'action' => 'leave'
+				)
+			);
+		}
+	}
+
+
+/**
+ * volunteer_leave
+ *
+ * @throws ForbiddenException
+ * @param string $id
+ * @return void
+*/
+	public function supervisor_leave($organization_id = null) 
+	{
+		if(_CurrentUserCanWrite($this->Auth->user('user_id')))
+		{
+			if ($organization_id != null) 
+			{
+				$user_id = $this->Auth->user('user_id');
+				$permission = $this->Organization->Permission->findByUserIdAndOrganizationId($user_id, $organization_id); // yes, this will work
+				if( empty($permission) )
+				{
+				// permissions didn't exist.  redirect gracefully
+					$this->Session->setFlash(__('We are unable to process your request at this time.'), 'danger');
+				}
+				$this->Organization->Permission->id = $permission['Permission']['permission_id'];
+				if( $this->Organization->Permission->saveField('write', false) )
+				{
+				// redirect
+					$this->Session->setFlash(__('You are no longer supervising events or users for this organization.'), 'success');
+				}
+				else
+				{
+				// didn't save
+					$this->Session->setFlash(__('We are unable to process your request at this time.'), 'danger');
+				}
+			}
+
+			$conditions = array(
+				'Organization.organization_id' => $this->_GetUserOrganizationsByPermission('publish')
+			);
+
+			$fields = array(
+				'Organization.*',
+			);
+
+			$this->Paginator->settings = array(
+				'conditions' => $conditions,
+				'limit' => 10,
+				'contain' => array()
+			);
+
+			$data = $this->Paginator->paginate();
+			$this->set(compact('data'));
+		}
+		else 
+		{
+			return $this->redirect(
+				array(
+					'coordinator' => true,
+					'controller' => 'organizations',
+					'action' => 'leave'
+				)
+			);
+		}
 	}
